services/graphql-gateway/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nuj-graphql-gateway"
version = "0.1.0"
description = "GraphQL API gateway for NUJ Monitor"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "strawberry-graphql[fastapi]>=0.216.0",
    "httpx>=0.25.0",
    "pydantic>=2.4.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]
